requirements.txt
transformers
datasets
pillow>=10.0.1
requests
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
matplotlib
tqdm
wandb
fastapi
python-multipart
uvicorn
